// contract/src/types/TodoList.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import type BN from "bn.js";
import type { ContractOptions } from "web3-eth-contract";
import type { EventLog } from "web3-core";
import type { EventEmitter } from "events";
import type {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface TodoList extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): TodoList;
  clone(): TodoList;
  methods: {
    completeTask(_id: number | string | BN): NonPayableTransactionObject<void>;

    createTask(_content: string): NonPayableTransactionObject<void>;

    getTask(
      _id: number | string | BN
    ): NonPayableTransactionObject<[string, string, boolean]>;

    taskCount(): NonPayableTransactionObject<string>;

    tasks(arg0: number | string | BN): NonPayableTransactionObject<{
      id: string;
      content: string;
      completed: boolean;
      0: string;
      1: string;
      2: boolean;
    }>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}